[core] Modify GKE to generate code to be  (#2043)
* [core] Modify GKE to generate code to be used exclusively with argument
conversion and argument synthesis.

This is the initial implementation to argument gathering on the C++
side. It changes the autogenerated C++ host entry point to gather the
arguments that are passed in and pass them as a std::vector<void*> to
the runtime for processing. The vector includes only textual arguments
for now. All hidden arguments are ignored.

More work will need to happen to support return values. Since no path is
actually returning arguments, not having this implemented immediately
won't change any behavior.

* Add layout test to verify that we agree with clang++ running under the
covers.