[core] Add support for kernel return values. (#2082)
* Add interface to exclude any permutation of arguments
when performing argument conversion.

There is both a "drop_front" style interface to support the legacy style
and a fully generalized interface to allow the filtering of any
permutation of arguments.

* Workaround bogus warning about doxygen tag.

* [core] Add support for kernel return values.

GKE has always supported return values from kernels. However, the
existing runtime intermediate layers do not use them and just drop them
on the floor, if they are present.

These patches fuse the old GKE with the new argument gathering approach
in a hybrid format. Specifically, GKE will construct both a data packet
with the argument data packed in it *and* a std::vector<void*> with
pointers to the argument values. A new runtime entry point can use
either strategy for the incoming direction. For the outgoing direction,
"synthesizing" the result value into the calling code is nonsensical and
an actual value must be returned, which was already supported by the old
altLaunchKernel protocol.