[JIT] Hook everything up to use the new argument synthesis (#2084)
* [JIT] Hook everything up to use the new argument synthesis

Add option to nvq++ to use the new kernel launcher.

Add code to the runtime to implement the new launch sequence. This
receives the new vector of arguments protocol, uses the new argument
conversion, and then calls the new argument synthesis pass to specialize
the kernel for JIT compilation.

Add a couple of tests to smoke test this new implementation.

* Add missing override.
Add ArgumentConversion module to the python library to resolve symbols.